InstaMemberRepository

+++ src/main/java/com/ll/gramgram/boundedContext/instaMember/repository/InstaMemberRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface InstaMemberRepository extends JpaRepository<InstaMember, Integer> {
+public interface InstaMemberRepository extends JpaRepository<InstaMember, Long> {
     Optional<InstaMember> findByUsername(String username);
 }
